convert ixml module to folder
